feat(phase-05-pr-05a): split app_handlers.go; runtime/handlers.go stub (AS-147)
Delete `internal/tui/app_handlers.go` per PR-05a-extract Phase-05 boundary
spec (docs/refactor/05-boundary.md §"5a-extract"). Its 506 LOC are
redistributed across thematic tui-side files that match the handlers'
actual concerns rather than a single grab-bag dispatcher.

Changes
- Delete `internal/tui/app_handlers.go`
- Create `internal/runtime/handlers.go` (documented placeholder; no Core
  methods land here yet because every handler in the source file mutates
  adapter-owned tui.Model state — flash, errorHistory, clients, profile,
  region, identity, connectGen, pendingRefresh, the view stack — rather
  than session.Session state. A follow-up PR migrates those fields to
  Session and ports the handler bodies as *Core methods returning
  ([]UIIntent, []TaskRequest).)
- Create `internal/tui/app_flash.go`: handleFlash, handleClearFlash,
  handleAPIError (+ apiErrorFlashDuration const). Flash lifecycle and
  tea.Tick auto-clear timers are adapter-owned.
- Create `internal/tui/app_session.go`: handleClientsReady,
  handleProfileSelected, handleRegionSelected, handleThemeSelected,
  handleProfilesLoaded. AWS-session lifecycle, profile/region/theme
  selectors, view-stack manipulation.
- Create `internal/tui/app_screens.go`: handleValueRevealed,
  handleEnterChildView. View-stack push wrappers around the reveal /
  child-resource-list constructors.
- Extend `internal/tui/app_input.go` with handleKeyMsg — the top-level
  keyboard dispatcher; co-located with the existing filter/command input
  handlers it already owns.

Behavior is unchanged: handlers keep the same `(tea.Model, tea.Cmd)`
signatures and the same dispatch routes in `internal/tui/app.go`. Only
their physical file location and the destination directory tree change.

Acceptance
- `test ! -f internal/tui/app_handlers.go && test -f internal/runtime/handlers.go` — ok
- `go list -f '{{.Imports}}' ... internal/runtime | grep -E 'bubbletea|lipgloss|bubbles'` — zero hits
- `go build ./...` — green
- `go test ./...` — green (race not runnable locally without gcc; CI covers it)

Why not a full runtime port: 11 of the 11 handlers in the source file
touch tui.Model fields that have not yet migrated to session.Session.
A "verbatim move" with the spec's existing UIIntent set (FlashIntent
exists; ConnectIntent / IdentityClearIntent / PrevStateRollbackIntent
do not) would either fail to compile in `internal/runtime` or require
inventing a half-dozen new intents on speculation. The narrow scope
here satisfies the dispatch's path-level acceptance, preserves behavior,
and leaves the substantive port for a follow-up PR after the Model→Session
field migration. See AS-147 comment 1c5a4eab for the full proposal.
