+++ internal/tui/app_session.go
@@ -1,197 +1,27 @@
 package tui
 
+// app_session.go — AWS-session lifecycle handlers: clients-ready bootstrap,
+// profile/region/theme selectors, and the profile-list loader. Split out of
+// internal/tui/app_handlers.go in Phase-05 PR-05a-h1 (AS-147). These handlers
+// mutate tui-Model fields (profile, region, identity, clients, connectGen,
+// pendingRefresh, hasPrevState, prevProfile, prevRegion, activeTheme,
+// headerCacheKey, preSuppliedClients, command, noCache) that have not yet
+// migrated to session.Session; once they do, these bodies move to *Core
+// methods returning ([]UIIntent, []TaskRequest) in a follow-up PR.
+
 import (
 	"fmt"
 	"os"
-	"strings"
 	"time"
 
-	"charm.land/bubbles/v2/key"
 	tea "charm.land/bubbletea/v2"
 
 	awsclient "github.com/k2m30/a9s/v3/internal/aws"
 	"github.com/k2m30/a9s/v3/internal/config"
-	"github.com/k2m30/a9s/v3/internal/resource"
 	"github.com/k2m30/a9s/v3/internal/tui/messages"
 	"github.com/k2m30/a9s/v3/internal/tui/styles"
 	"github.com/k2m30/a9s/v3/internal/tui/views"
 )
-
-// apiErrorFlashDuration controls how long API error messages stay visible.
-// Longer than regular flash (2s) because error messages are more important.
-const apiErrorFlashDuration = 5 * time.Second
-
-// handleKeyMsg processes all keyboard input: force-quit, input modes, global
-// keys, then falls through to the active view.
-func (m Model) handleKeyMsg(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	// Global keys handled before delegation
-	if key.Matches(msg, m.keys.ForceQuit) {
-		return m, tea.Quit
-	}
-
-	// Handle input modes — don't clear error hint during text input.
-	switch m.inputMode {
-	case modeFilter:
-		return m.updateFilterMode(msg)
-	case modeCommand:
-		return m.updateCommandMode(msg)
-	}
-
-	// Clear error hint on any navigation keypress (not during text input).
-	m.showErrorHint = false
-
-	// If the active view is in search input mode, delegate all keys to it.
-	// This prevents global keys (q, ?, i, etc.) from firing while typing a search query.
-	if s, ok := m.activeView().(views.Searchable); ok && s.IsSearchInputMode() {
-		return m.updateActiveView(msg)
-	}
-
-	// Global keys in normal mode
-	if key.Matches(msg, m.keys.Help) {
-		// If already on help, let the help view handle it (closes help)
-		if _, ok := m.activeView().(*views.HelpModel); ok {
-			return m.updateActiveView(msg)
-		}
-		ctx := m.helpContext()
-		activeShortName := ""
-		if rl, ok := m.activeView().(*views.ResourceListModel); ok {
-			activeShortName = rl.ShortName()
-		}
-		help := views.NewHelpWithResource(m.keys, ctx, activeShortName)
-		help.SetSize(m.innerSize())
-		m.pushView(&help)
-		return m, nil
-	}
-	if key.Matches(msg, m.keys.Identity) {
-		// If already on identity view, let it handle the key (dismisses)
-		if _, ok := m.activeView().(*views.IdentityModel); ok {
-			return m.updateActiveView(msg)
-		}
-		id := views.NewIdentity(m.profile, m.region, m.keys)
-		if m.identity != nil {
-			data := m.identityToViewData()
-			id.SetIdentity(data)
-		}
-		id.SetSize(m.innerSize())
-		m.pushView(&id)
-		// Always re-fetch on i press
-		m.identityFetching = true
-		cmd := m.fetchIdentity()
-		return m, cmd
-	}
-	if key.Matches(msg, m.keys.ErrorLog) {
-		// If already viewing the error log, let the view handle the key.
-		if ym, ok := m.activeView().(*views.YAMLModel); ok && ym.IsTextViewer() {
-			return m.updateActiveView(msg)
-		}
-		if len(m.errorHistory) == 0 {
-			return m.handleFlash(messages.FlashMsg{Text: "No errors this session"})
-		}
-		var sb strings.Builder
-		for i := len(m.errorHistory) - 1; i >= 0; i-- {
-			e := m.errorHistory[i]
-			fmt.Fprintf(&sb, "[%s] %s\n", e.time.Format("15:04:05"), e.message)
-		}
-		tv := views.NewTextViewer("errors", sb.String(), m.keys)
-		tv.SetSize(m.innerSize())
-		m.pushView(&tv)
-		return m, nil
-	}
-	if key.Matches(msg, m.keys.Quit) {
-		return m, tea.Quit
-	}
-	if key.Matches(msg, m.keys.Escape) {
-		if d, ok := m.activeView().(*views.DetailModel); ok && d.ConsumesEscapeLocally() {
-			return m.updateActiveView(msg)
-		}
-		// If active view has active search (confirmed highlights), delegate Esc to clear it.
-		if s, ok := m.activeView().(views.Searchable); ok && s.IsSearchActive() {
-			return m.updateActiveView(msg)
-		}
-		// Related-navigation resource lists should pop immediately on Esc.
-		if rl, ok := m.activeView().(*views.ResourceListModel); ok && rl.EscPops() {
-			m.popView()
-			return m, nil
-		}
-		// If active view has a confirmed filter, clear it first
-		if f, ok := m.activeView().(views.Filterable); ok && f.GetFilter() != "" {
-			f.SetFilter("")
-			if rl, ok := m.activeView().(*views.ResourceListModel); ok {
-				m.cacheTopLevelResourceList(*rl)
-			}
-			return m, nil
-		}
-		// Otherwise pop view; no-op on main menu (never quit from Esc)
-		m.popView()
-		return m, nil
-	}
-	if key.Matches(msg, m.keys.Colon) {
-		m.inputMode = modeCommand
-		m.cmdInput.Reset()
-		m.cmdInput.Focus()
-		return m, nil
-	}
-	if key.Matches(msg, m.keys.Filter) {
-		// Only activate filter mode on filterable views
-		if _, ok := m.activeView().(views.Filterable); ok {
-			m.inputMode = modeFilter
-			m.cmdInput.Reset()
-			m.cmdInput.Focus()
-			return m, nil
-		}
-		// On help views, delegate / to the view (which sends PopViewMsg to close).
-		if _, ok := m.activeView().(*views.HelpModel); ok {
-			return m.updateActiveView(msg)
-		}
-		// On searchable views (detail, YAML), delegate / for search activation.
-		if _, ok := m.activeView().(views.Searchable); ok {
-			return m.updateActiveView(msg)
-		}
-		// On other static views (reveal), consume / without action.
-		return m, nil
-	}
-
-	// Copy (c) — context-dependent clipboard copy
-	if key.Matches(msg, m.keys.Copy) {
-		return m.handleCopy()
-	}
-
-	// Refresh (ctrl+r) — re-fetch resources in resource list
-	if key.Matches(msg, m.keys.Refresh) {
-		return m.handleRefresh()
-	}
-
-	// Reveal (x) — fetch and display value via registered reveal fetcher
-	if key.Matches(msg, m.keys.Reveal) {
-		return m.handleReveal()
-	}
-
-	return m.updateActiveView(msg)
-}
-
-// handleFlash sets the flash message and schedules its auto-clear.
-func (m Model) handleFlash(msg messages.FlashMsg) (tea.Model, tea.Cmd) {
-	newGen := m.flash.gen + 1
-	m.flash = flashState{text: msg.Text, isError: msg.IsError, active: true, gen: newGen}
-	if msg.IsError {
-		m.errorHistory = append(m.errorHistory, errorEntry{time: time.Now(), message: msg.Text})
-	}
-	gen := m.flash.gen
-	return m, tea.Tick(2*time.Second, func(_ time.Time) tea.Msg {
-		return messages.ClearFlashMsg{Gen: gen}
-	})
-}
-
-// handleClearFlash clears the flash if the generation matches (not stale).
-func (m Model) handleClearFlash(msg messages.ClearFlashMsg) (tea.Model, tea.Cmd) {
-	if msg.Gen == m.flash.gen {
-		if m.flash.isError {
-			m.showErrorHint = true
-		}
-		m.flash.active = false
-	}
-	return m, nil
-}
 
 // handleClientsReady stores the new AWS clients and optionally triggers a
 // pending refresh (after profile/region switch).
@@ -451,56 +281,4 @@ func (m Model) handleProfilesLoaded(msg profilesLoadedMsg) (tea.Model, tea.Cmd) 
 	p.SetSize(m.innerSize())
 	m.pushView(&p)
 	return m, nil
-}
-
-// handleValueRevealed pushes the reveal view or flashes an error.
-func (m Model) handleValueRevealed(msg messages.ValueRevealedMsg) (tea.Model, tea.Cmd) {
-	if msg.Err != nil {
-		errText := "reveal failed: " + msg.Err.Error()
-		return m, func() tea.Msg {
-			return messages.FlashMsg{Text: errText, IsError: true}
-		}
-	}
-	rv := views.NewReveal(msg.ResourceID, msg.Value, m.keys)
-	rv.SetSize(m.innerSize())
-	m.pushView(&rv)
-	return m, nil
-}
-
-// handleEnterChildView creates a child resource list view and kicks off a fetch
-// using the child type registry. This is the generic handler for all child views.
-func (m Model) handleEnterChildView(msg messages.EnterChildViewMsg) (tea.Model, tea.Cmd) {
-	childTypeDef := resource.GetChildType(msg.ChildType)
-	if childTypeDef == nil {
-		return m, func() tea.Msg {
-			return messages.FlashMsg{Text: fmt.Sprintf("unknown child type: %s", msg.ChildType), IsError: true}
-		}
-	}
-	rl := views.NewChildResourceList(*childTypeDef, msg.ParentContext, msg.DisplayName, m.viewConfig, m.keys)
-	rl.SetSize(m.innerSize())
-	rl, initCmd := rl.Init()
-	m.pushView(&rl)
-	fetchCmd := m.fetchChildResources(msg.ChildType, msg.ParentContext)
-	return m, tea.Batch(initCmd, fetchCmd)
-}
-
-// handleAPIError shows a flash error and clears loading state on the resource list.
-func (m Model) handleAPIError(msg messages.APIErrorMsg) (tea.Model, tea.Cmd) {
-	code, message, _ := awsclient.ClassifyAWSError(msg.Err)
-	var flashText string
-	if code != "" && code != "Unknown" {
-		flashText = fmt.Sprintf("[%s] %s", code, message)
-	} else {
-		flashText = msg.Err.Error()
-	}
-	newGen := m.flash.gen + 1
-	m.flash = flashState{text: flashText, isError: true, active: true, gen: newGen}
-	m.errorHistory = append(m.errorHistory, errorEntry{time: time.Now(), message: flashText})
-	if rl, ok := m.activeView().(*views.ResourceListModel); ok {
-		rl.ClearLoading()
-	}
-	gen := m.flash.gen
-	return m, tea.Tick(apiErrorFlashDuration, func(_ time.Time) tea.Msg {
-		return messages.ClearFlashMsg{Gen: gen}
-	})
 }
